# crates/public-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "canopy public-server",
    "description": "Internet-facing API for the canopy fleet. Device-authenticated endpoints require an mTLS client certificate.\n\nRequest bodies may be sent compressed, and this is recommended for large payloads: the server transparently decodes `Content-Encoding: gzip`, `br`, `deflate`, or `zstd`. The accepted encodings are also listed structurally in the `x-request-compression` extension.",
    "contact": {
      "name": "BES Developers",
      "email": "contact@bes.au"
    },
    "license": {
      "name": "GPL-3.0-or-later"
    },
    "version": ""
  },
  "paths": {
...
    "/restore-verification": {
      "post": {
        "tags": [
          "restore"
        ],
        "summary": "Report the outcome of a restore attempt and the replica's health.",
        "description": "Records a verification report for a restore the device performed from its\nworklist. Send one report per attempt, on success and on failure alike. A\nreport with a `success` outcome and `replica_healthy: true` marks the\nsnapshot as verified; for run-once intents this is what removes the entry\nfrom `GET /restore-worklist` until a newer snapshot appears.\n\nAuthorization matches `POST /restore-credentials`: the device must hold an\nenabled restore declaration covering the reported group and type,\notherwise the request is rejected with 403.\n\nThe report names the declaration it is about, and that declaration must\nstill exist and belong to the calling consumer. A replica nothing declares\nany more is not one Canopy tracks, so a report naming a retired declaration\nis refused rather than recorded against a replica that could never recover.",
        "operationId": "verification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Report recorded."
          },
          "403": {
            "description": "No enabled declaration authorizes this (group, type), or the named declaration belongs to another consumer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "404": {
            "description": "The named declaration does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "backup-restore-device": []
          }
        ]
      }
    },
...
    "/servers": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List publicly-listed central applications.",
        "description": "Returns every central server that has both a public display name and a\nreachable host configured, ordered by environment tier (production\nfirst, then clone, demo, test, dev) and then by name. Used by clients\nto let a user pick which server to connect to.",
        "operationId": "list_servers",
        "responses": {
          "200": {
            "description": "Publicly-listed central applications, ordered by rank then name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicServer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/servers/register/begin": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Start device enrollment against a server.",
        "description": "Validates the enrollment token against the given server and, if valid,\nissues a short-lived (5 minute) signed challenge bound to the server\nID, the token, and the caller's public key. The device must sign this\nchallenge and submit it to the completion endpoint to finish\nenrollment; the token itself is validated here but not yet consumed.\n\nThis endpoint is rate-limited per source IP and per target server; a\ntripped limit returns 429. Any other failure — an unknown or deleted\nserver, or an invalid or expired token — is surfaced as a generic 403,\ndeliberately not distinguishing which check failed.",
        "operationId": "register_begin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/servers/register/complete": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Complete device enrollment by presenting a signed challenge.",
        "description": "Verifies the signature over the challenge transcript using the public\nkey supplied here, then binds the device to the server: an existing\ndevice re-enrolling with the same key is reused as-is; a device\nre-enrolling with a different key replaces the server's previous\ndevice (revoking that device's access); otherwise a new device\nidentity is created. On success the device is granted the server\nrole, the enrollment token is consumed, and the server is marked as\nregistered.\n\nEnrollment is refused if the presented public key is already bound to\na different live server. Like the start-enrollment endpoint, this one\nis rate-limited per source IP and per target server (429 on a tripped\nlimit) and reports every other kind of failure as a generic 403.",
        "operationId": "register_complete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/servers/self": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Report the calling device's own identity.",
        "description": "Resolves the caller from its device certificate and returns the server\nit is enrolled as together with its own device ID — the same pair\nreturned when the device completed enrollment. A device authenticates\nentirely from its certificate, so it never needs these IDs to make\ncalls; this endpoint lets one that has lost track of them recover them.\n\n- **401**: the request has no client certificate, or the certificate\n  doesn't match a known device.\n- **409**: the calling device is attached to more than one server, which\n  should not normally happen; contact support if you see this.\n- **412**: the device is registered but has not yet been attached to a\n  server.",
        "operationId": "server_self",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          },
          "412": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "server-device": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PublicServer": {
        "type": "object",
        "description": "A publicly-listed central server that a client can connect to.",
        "required": [
          "name",
          "host"
        ],
        "properties": {
          "host": {
            "$ref": "#/components/schemas/UrlField",
            "description": "The server's reachable base URL."
          },
          "name": {
            "type": "string",
            "description": "Public-facing display name of the server."
          },
          "rank": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerRank",
                "description": "The server's environment tier (production, clone, demo, test, or\ndev), if set. Used to order the listing and to let clients label\nnon-production entries."
              }
            ]
          }
        }
      },
...
      "UrlField": {
        "type": "string",
        "format": "uri",
        "description": "A URL, given as a plain string. Any trailing slash is stripped when the\nvalue is returned."
      },
      "VerificationArgs": {
        "type": "object",
        "description": "Report of a restore attempt and the health of the resulting replica.",
        "required": [
          "replica_id",
          "group",
          "server_id",
          "type",
          "intent",
          "outcome",
          "replica_healthy",
          "observed_at"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable error detail, when the restore failed."
          },
          "group": {
            "type": "string",
            "format": "uuid",
            "description": "The server group whose backup was restored."
          },
          "health_details": {
            "description": "Arbitrary structured health data to record alongside the report\n(database statistics, whether indexes needed rebuilding, and so on).\nStored and displayed as-is."
          },
          "intent": {
            "type": "string",
            "description": "The restore intent this attempt was performed under."
          },
          "migration": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MigrationArgs",
                "description": "What the migrations did, for a report under a `migrate` intent. Omit for\nevery other intent."
              }
            ]
          },
          "observed_at": {
            "type": "string",
            "description": "When the restore result was observed, as an RFC 3339 timestamp."
          },
          "outcome": {
            "type": "string",
            "description": "Whether the restore succeeded (`success`) or failed (`failure`)."
          },
          "postgres_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Version of the PostgreSQL server the data was restored into, if\napplicable."
          },
          "redaction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RedactionArgs",
                "description": "What the masking manifest did, for a replica that redacts. Omit for a\nreplica that doesn't."
              }
            ]
          },
          "replica_healthy": {
            "type": "boolean",
            "description": "Whether the restored database came up healthy and passed readiness\nchecks. A replica only counts as verified when the outcome is\n`success` and this is `true`."
          },
          "replica_id": {
            "type": "string",
            "format": "uuid",
            "description": "The declaration this report concerns, taken from the worklist entry's\n`replica_id`. Required: several replicas can share one group, server,\ntype, and intent, so a report that named no declaration could not be\nattributed to one of them."
          },
          "run_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "This must be the run-uuid the client minted for this run.\nThe field is optional only so older clients don't break; it WILL be made\nmandatory in future."
          },
          "s3_received_payload_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes of decoded object payload received from S3 during the restore."
          },
          "s3_received_raw_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes of raw HTTP traffic received from S3 during the restore,\nincluding protocol overhead."
          },
          "s3_sent_payload_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes of decoded object payload sent to S3 during the restore."
          },
          "s3_sent_raw_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes of raw HTTP traffic sent to S3 during the restore, including\nprotocol and signing overhead. Omit when traffic was not measured."
          },
          "server_id": {
            "type": "string",
            "format": "uuid",
            "description": "The server whose backup was restored."
          },
          "snapshot_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Identifier of the snapshot that was restored. Omit on a failure that\nnever got as far as selecting a snapshot."
          },
          "type": {
            "type": "string",
            "description": "The backup type that was restored (e.g. `tamanu-postgres`)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `admin` role."
      },
      "backup-restore-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `backup-restore` role (or `admin`)."
      },
      "releaser-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `releaser` role (or `admin`)."
      },
      "server-device": {
        "type": "mutualTLS",
        "description": "mTLS client certificate for a device with the `server` role (or `admin`)."
      }
    }
  },
...
}